Move filter utils to utils
