+++ server/src/core/api-rest/api-rest.service.ts
@@ -14,28 +14,15 @@ import { EnvironmentService } from 'src/integrations/environment/environment.ser
 import { ObjectMetadataService } from 'src/metadata/object-metadata/object-metadata.service';
 import { capitalize } from 'src/utils/capitalize';
 import { mapFieldMetadataToGraphqlQuery } from 'src/core/api-rest/utils/map-field-metadata-to-graphql-query.utils';
+import {
+  addDefaultConjunctionIfMissing,
+  checkFilterQuery,
+  parseStringFilter,
+} from 'src/core/api-rest/utils/filter-query.utils';
+import { checkFields } from 'src/core/api-rest/utils/metadata-query.utils';
 
-enum FILTER_COMPARATORS {
-  eq = 'eq',
-  neq = 'neq',
-  in = 'in',
-  is = 'is',
-  gt = 'gt',
-  gte = 'gte',
-  lt = 'lt',
-  lte = 'lte',
-  startsWith = 'startsWith',
-  like = 'like',
-  ilike = 'ilike',
-}
 const ALLOWED_DEPTH_VALUES = [1, 2];
 const DEFAULT_ORDER_DIRECTION = OrderByDirection.AscNullsFirst;
-enum CONJUNCTIONS {
-  or = 'or',
-  and = 'and',
-  not = 'not',
-}
-const DEFAULT_FILTER_CONJUNCTION = CONJUNCTIONS.and;
 
 @Injectable()
 export class ApiRestService {
@@ -171,187 +158,6 @@ export class ApiRestService {
     };
   }
 
-  addDefaultConjunctionIfMissing(filterQuery) {
-    if (!(filterQuery.includes('(') && filterQuery.includes(')'))) {
-      return `${DEFAULT_FILTER_CONJUNCTION}(${filterQuery})`;
-    }
-    return filterQuery;
-  }
-
-  checkFilterQuery(filterQuery) {
-    const countOpenedBrackets = (filterQuery.match(/\(/g) || []).length;
-    const countClosedBrackets = (filterQuery.match(/\)/g) || []).length;
-    const diff = countOpenedBrackets - countClosedBrackets;
-    if (diff !== 0) {
-      const hint =
-        diff > 0
-          ? `${diff} open bracket${diff > 1 ? 's are' : ' is'}`
-          : `${Math.abs(diff)} close bracket${
-              Math.abs(diff) > 1 ? 's are' : ' is'
-            }`;
-      throw Error(`'filter' invalid. ${hint} missing in the query`);
-    }
-    return;
-  }
-
-  parseFilterQueryContent(filterQuery) {
-    let bracketsCounter = 0;
-    let parenthesisCounter = 0;
-    const predicates: string[] = [];
-    let currentPredicates = '';
-    for (const c of filterQuery) {
-      if (c === '(') {
-        parenthesisCounter++;
-        if (parenthesisCounter === 1) continue;
-      }
-      if (c === ')') {
-        parenthesisCounter--;
-        if (parenthesisCounter === 0) continue;
-      }
-      if (c === '[') {
-        bracketsCounter++;
-      }
-      if (c === ']') {
-        bracketsCounter--;
-      }
-      if (c === ',' && parenthesisCounter === 1 && bracketsCounter === 0) {
-        predicates.push(currentPredicates);
-        currentPredicates = '';
-        continue;
-      }
-      if (parenthesisCounter >= 1) currentPredicates += c;
-    }
-    if (currentPredicates.length) {
-      predicates.push(currentPredicates);
-    }
-    return predicates;
-  }
-
-  parseStringFilter(filterQuery, objectMetadataItem) {
-    const result = {};
-    const match = filterQuery.match(
-      `^(${Object.values(CONJUNCTIONS).join('|')})((.+))$`,
-    );
-    if (match) {
-      const conjunction = match[1];
-      const subResult = this.parseFilterQueryContent(filterQuery).map((elem) =>
-        this.parseStringFilter(elem, objectMetadataItem),
-      );
-      if (conjunction === CONJUNCTIONS.not) {
-        if (subResult.length > 1) {
-          throw Error(
-            `'filter' invalid. 'not' conjunction should contain only 1 condition. eg: not(field[eq]:1)`,
-          );
-        }
-        result[conjunction] = subResult[0];
-      } else {
-        result[conjunction] = subResult;
-      }
-      return result;
-    }
-    return this.parseSimpleFilter(filterQuery, objectMetadataItem);
-  }
-
-  parseSimpleFilterString(filterString: string): {
-    fields: string[];
-    comparator: string;
-    value: string;
-  } {
-    if (
-      !filterString.match(
-        `^(.+)\\[(${Object.keys(FILTER_COMPARATORS).join('|')})\\]:(.+)$`,
-      )
-    ) {
-      throw Error(`'filter' invalid for '${filterString}'. eg: price[gte]:10`);
-    }
-    let fields = '';
-    let comparator = '';
-    let value = '';
-    let previousC = '';
-    let fillFields = true;
-    let fillComparator = false;
-    let fillValue = false;
-    for (const c of filterString) {
-      if (c === '[') {
-        fillFields = false;
-      }
-      if (previousC === '[' && !comparator.length) {
-        fillComparator = true;
-      }
-      if (c === ']') {
-        fillComparator = false;
-      }
-      if (previousC === ']' && c === ':' && !value.length) {
-        fillValue = true;
-        continue;
-      }
-      if (fillFields) {
-        fields += c;
-      }
-      if (fillComparator) {
-        comparator += c;
-      }
-      if (fillValue) {
-        value += c;
-      }
-      previousC = c;
-    }
-    if (!Object.keys(FILTER_COMPARATORS).includes(comparator)) {
-      throw Error(
-        `'filter' invalid for '${filterString}', comparator ${comparator} not in ${Object.keys(
-          FILTER_COMPARATORS,
-        ).join(',')}`,
-      );
-    }
-    return { fields: fields.split('.'), comparator, value };
-  }
-
-  parseSimpleFilter(filterString: string, objectMetadataItem) {
-    // price[lte]:100
-    const { fields, comparator, value } =
-      this.parseSimpleFilterString(filterString);
-    this.checkFields(objectMetadataItem, fields, 'filter');
-    const fieldType = this.getFieldType(objectMetadataItem, fields[0]);
-    const formattedValue = this.formatFieldValue(value, fieldType, comparator);
-    return fields.reverse().reduce(
-      (acc, currentValue) => {
-        return { [currentValue]: acc };
-      },
-      { [comparator]: formattedValue },
-    );
-  }
-
-  formatFieldValue(value, fieldType?, comparator?) {
-    if (comparator === 'in') {
-      if (value[0] !== '[' || value[value.length - 1] !== ']') {
-        throw Error(
-          `'filter' invalid for 'in' operator. Received '${value}' but array value expected eg: 'field[in]:[value_1,value_2]'`,
-        );
-      }
-      const stringValues = value.substring(1, value.length - 1);
-      return stringValues
-        .split(',')
-        .map((value) => this.formatFieldValue(value, fieldType));
-    }
-    if (comparator === 'is') {
-      return value;
-    }
-    if (fieldType === 'NUMBER') {
-      return parseInt(value);
-    }
-    if (fieldType === 'BOOLEAN') {
-      return value.toLowerCase() === 'true';
-    }
-    if (
-      (value[0] === '"' || value[0] === "'") &&
-      (value.charAt(value.length - 1) === '"' ||
-        value.charAt(value.length - 1) === "'")
-    ) {
-      return value.substring(1, value.length - 1);
-    }
-    return value;
-  }
-
   parseFilter(request, objectMetadataItem) {
     const parsedObjectId = this.parseObject(request)[1];
     if (parsedObjectId) {
@@ -363,9 +169,9 @@ export class ApiRestService {
     if (typeof rawFilterQuery !== 'string') {
       return {};
     }
-    this.checkFilterQuery(rawFilterQuery);
-    const filterQuery = this.addDefaultConjunctionIfMissing(rawFilterQuery);
-    return this.parseStringFilter(filterQuery, objectMetadataItem);
+    checkFilterQuery(rawFilterQuery);
+    const filterQuery = addDefaultConjunctionIfMissing(rawFilterQuery);
+    return parseStringFilter(filterQuery, objectMetadataItem);
   }
 
   parseOrderBy(request, objectMetadataItem) {
@@ -396,37 +202,8 @@ export class ApiRestService {
         result[orderByItem] = DEFAULT_ORDER_DIRECTION;
       }
     }
-    this.checkFields(objectMetadataItem, Object.keys(result), 'order_by');
+    checkFields(objectMetadataItem, Object.keys(result));
     return <RecordOrderBy>result;
-  }
-
-  checkFields(objectMetadataItem, fieldNames, queryParam) {
-    for (const fieldName of fieldNames) {
-      if (
-        !objectMetadataItem.fields
-          .reduce(
-            (acc, itemField) => [
-              ...acc,
-              itemField.name,
-              ...Object.keys(itemField.targetColumnMap),
-            ],
-            [],
-          )
-          .includes(fieldName)
-      ) {
-        throw Error(
-          `'${queryParam}' field '${fieldName}' does not exist in '${objectMetadataItem.targetTableName}' object`,
-        );
-      }
-    }
-  }
-
-  getFieldType(objectMetadataItem, fieldName) {
-    for (const itemField of objectMetadataItem.fields) {
-      if (fieldName === itemField.name) {
-        return itemField.type;
-      }
-    }
   }
 
   parseLimit(request) {
