+++ server/src/core/api-rest/utils/__tests__/map-field-metadata-to-graphql-query.spec.ts
@@ -1,36 +1,20 @@
 import { mapFieldMetadataToGraphqlQuery } from 'src/core/api-rest/utils/map-field-metadata-to-graphql-query.utils';
+import {
+  fieldCurrency,
+  fieldLink,
+  fieldNumber,
+  fieldString,
+  objectMetadataItem,
+} from 'src/core/api-rest/utils/__tests__/utils';
 
 describe('mapFieldMetadataToGraphqlQuery', () => {
-  const fieldNumber = {
-    name: 'fieldNumber',
-    type: 'NUMBER',
-    targetColumnMap: { value: 'fieldNumber' },
-  };
-  const fieldString = {
-    name: 'fieldString',
-    type: 'TEXT',
-    targetColumnMap: { value: 'fieldString' },
-  };
-  const fieldLink = {
-    name: 'fieldLink',
-    type: 'LINK',
-    targetColumnMap: { label: 'fieldLinkLabel', url: 'fieldLinkUrl' },
-  };
-  const fieldCurrency = {
-    name: 'fieldCurrency',
-    type: 'CURRENCY',
-    targetColumnMap: {
-      amountMicros: 'fieldCurrencyAmountMicros',
-      currencyCode: 'fieldCurrencyCurrencyCode',
-    },
-  };
-  const objectMetadataItem = {
-    fields: [fieldNumber, fieldString, fieldLink],
-  };
   it('should map properly', () => {
     expect(
       mapFieldMetadataToGraphqlQuery(objectMetadataItem, fieldNumber),
     ).toEqual('fieldNumber');
+    expect(
+      mapFieldMetadataToGraphqlQuery(objectMetadataItem, fieldString),
+    ).toEqual('fieldString');
     expect(mapFieldMetadataToGraphqlQuery(objectMetadataItem, fieldLink))
       .toEqual(`
       fieldLink
