+++ server/src/core/api-rest/utils/__tests__/utils.ts
@@ -1,0 +1,26 @@
+export const fieldNumber = {
+  name: 'fieldNumber',
+  type: 'NUMBER',
+  targetColumnMap: { value: 'fieldNumber' },
+};
+export const fieldString = {
+  name: 'fieldString',
+  type: 'TEXT',
+  targetColumnMap: { value: 'fieldString' },
+};
+export const fieldLink = {
+  name: 'fieldLink',
+  type: 'LINK',
+  targetColumnMap: { label: 'fieldLinkLabel', url: 'fieldLinkUrl' },
+};
+export const fieldCurrency = {
+  name: 'fieldCurrency',
+  type: 'CURRENCY',
+  targetColumnMap: {
+    amountMicros: 'fieldCurrencyAmountMicros',
+    currencyCode: 'fieldCurrencyCurrencyCode',
+  },
+};
+export const objectMetadataItem = {
+  fields: [fieldNumber, fieldString, fieldLink],
+};
